Feature/fix service ms usuario catalgo stock

+++ ms-usuarios/src/main/java/cl/pchardware/usuarios/dto/RolResponse.java
@@ -4,8 +4,7 @@ import lombok.Data;
 
 @Data
 public class RolResponse {
-
-    private Integer idRol;
+    private Long idRol;
     private String nombre;
     private String descripcion;
 }
